requirements.txt
paho-mqtt==2.1.0
influxdb-client[async]==1.47.0
python-dotenv==1.0.1
